Added Docker Services to the project
[+] Added docker service for postgres and elasticsearch

[*] Moved application to inside the service->web
