+++ .env.dev-sample
@@ -1,4 +1,16 @@
+FLASK_ENV=development
+FLASK_APP=microblog.py
+PYTHONPATH=/home/microblog
 SECRET_KEY=a-really-long-and-unique-key-that-nobody-knows
+
 MAIL_SERVER=localhost
 MAIL_PORT=25
-ELASTICSEARCH_URL=http://localhost:9200
+
+DB_HOST=db
+DB_PORT=5432
+DB_USER=microblog_user
+DB_PASSWORD=password123
+DB_NAME=microblog
+
+SEARCH_HOST=es01
+SEARCH_PORT=9200
