--- .flaskenv
@@ -1,2 +1,0 @@
-FLASK_ENV=development
-FLASK_APP=microblog.py
