+++ services/web/app/auth/templates/registration.html
