+++ services/web/tests.py
@@ -1,9 +1,9 @@
 from datetime import datetime, timedelta
 import unittest
 
-from app import create_app, db
-from app.models import User, Post
-from config import Config
+from services.web.app import create_app, db
+from services.web.app.models import User, Post
+from services.web.config import Config
 
 
 class TestConfig(Config):
